packages/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0
supabase==2.3.0
python-dotenv>=1.0.0
mcp>=1.0.0
